First test for PgCtlWrapper
Created a class, PgCtlWrapper, that will allow to invoke pg_ctl with the desired
arguments. This class is being tested individually, in order to ensure that its
calls to the different commands in pg_ctl (start, stop, status and restart) are
working correctly.
It looked interesting to start the testing and building of the class by ensuring
that it detects the cases in which it cannot run pg_ctl start. Given that it is
run via ProcessBuilder, the errors can come by two ways:
-stdOut: pg_ctl itself only logs its output to stdout, even its errors. So any
error will be logged here. The chosen way to test this is by including another
cluster whose pg_hba.conf file is wrong.
-stdErr: running a command line command can also lead to the OS logging errors
to stdOut. In this case, the test points to an empty folder as cluster folder.

These two tests are working. The normal pg_ctl start test is ignored until
pg_ctl stop is implemented, in order not to leave an open cluster.