Creating autogeneratable maven project
Created module pgsupplier. This module will be able to auto generate
itself, and assemble itself into a jar. It needs some properties to
be set by the user: postgres.major, postgres.minor, postgres.os and
postgres.arch. So the command would be something with the pattern:
mvn clean generate-sources package -Dx=y -Dx=y...
The part of generating the target is complete, the next task is to
assemble the jar.